[CONTINUUM-1777] Able to create a build template without build definition Submitted by: John Michael Luy (with modifications) Submitted by: Gwen Harold Autencio (with modifications)
git-svn-id: https://svn.apache.org/repos/asf/continuum/trunk@1491782 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-webapp/src/main/java/org/apache/maven/continuum/web/action/admin/BuildDefinitionTemplateAction.java
@@ -206,7 +206,7 @@ public class BuildDefinitionTemplateAction
         return SUCCESS;
     }
 
-    private List<BuildDefinition> getBuildDefinitionsFromSelectedBuildDefinitions()
+    public List<BuildDefinition> getBuildDefinitionsFromSelectedBuildDefinitions()
         throws ContinuumException
     {
         if ( this.selectedBuildDefinitionIds == null )
@@ -293,9 +293,17 @@ public class BuildDefinitionTemplateAction
     {
         if ( confirmed )
         {
-            buildDefinition = getContinuum().getBuildDefinitionService().getBuildDefinition(
-                this.buildDefinition.getId() );
-            this.getContinuum().getBuildDefinitionService().removeBuildDefinition( buildDefinition );
+            if ( getContinuum().getBuildDefinitionService().isBuildDefinitionInUse( buildDefinition ) )
+            {
+                addActionError( getText( "buildDefinition.used" ) );
+                return ERROR;
+            }
+            else
+            {
+                buildDefinition =
+                    getContinuum().getBuildDefinitionService().getBuildDefinition( this.buildDefinition.getId() );
+                this.getContinuum().getBuildDefinitionService().removeBuildDefinition( buildDefinition );
+            }
         }
         else
         {
